Use NuGet.ProjectModel for compilation deps
