+++ src/Dotnet.Script.DependencyModel/Compilation/CompilationDependency.cs
@@ -1,0 +1,28 @@
+using System.Collections.Generic;
+
+namespace Dotnet.Script.DependencyModel.Compilation
+{
+    public class CompilationDependency
+    {
+        public CompilationDependency(string name, string version, IReadOnlyList<string> assemblyPaths, IReadOnlyList<string> scripts)
+        {
+            Name = name;
+            Version = version;
+            AssemblyPaths = assemblyPaths;
+            Scripts = scripts;
+        }
+
+        public string Name { get; }
+
+        public string Version { get; }
+
+        public IReadOnlyList<string> AssemblyPaths { get; }
+
+        public IReadOnlyList<string> Scripts { get; }
+
+        public override string ToString()
+        {
+            return $"Name: {Name} , Version: {Version}";
+        }
+    }
+}
