+++ src/Dotnet.Script.DependencyModel/Compilation/CompilationDependencyResolver.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -6,152 +6,52 @@ using Dotnet.Script.DependencyModel.Context;
 using Dotnet.Script.DependencyModel.Logging;
 using Dotnet.Script.DependencyModel.Process;
 using Dotnet.Script.DependencyModel.ProjectSystem;
-using Dotnet.Script.DependencyModel.ScriptPackage;
-using Microsoft.Extensions.DependencyModel;
-using Microsoft.Extensions.DependencyModel.Resolution;
-using NuGet.Packaging;
 
 namespace Dotnet.Script.DependencyModel.Compilation
 {
     public class CompilationDependencyResolver
     {
-        private readonly ScriptProjectProvider _scriptProjectProvider;
-        private readonly ScriptDependencyInfoProvider _scriptDependencyInfoProvider;
-        private readonly ScriptFilesDependencyResolver _scriptFilesDependencyResolver;
         private readonly Logger _logger;
+        private readonly ScriptProjectProvider _scriptProjectProvider;
+        private readonly ScriptDependencyContextReader _scriptDependencyContextReader;
+        private readonly IRestorer _restorer;
 
-        private CompilationDependencyResolver(ScriptProjectProvider scriptProjectProvider, ScriptDependencyInfoProvider scriptDependencyInfoProvider, ScriptFilesDependencyResolver scriptFilesDependencyResolver,  LogFactory logFactory)
+        public CompilationDependencyResolver(ScriptProjectProvider scriptProjectProvider, ScriptDependencyContextReader scriptDependencyContextReader, LogFactory logFactory)
         {
             _scriptProjectProvider = scriptProjectProvider;
-            _scriptDependencyInfoProvider = scriptDependencyInfoProvider;
-            _scriptFilesDependencyResolver = scriptFilesDependencyResolver;
-            _logger = logFactory.CreateLogger<CompilationDependencyResolver>();
+            _scriptDependencyContextReader = scriptDependencyContextReader;
+            _restorer = CreateRestorer(logFactory);
         }
 
         public CompilationDependencyResolver(LogFactory logFactory)
             : this
             (
                 new ScriptProjectProvider(logFactory),
-                new ScriptDependencyInfoProvider(CreateRestorer(logFactory), logFactory),
-                new ScriptFilesDependencyResolver(logFactory),
+                new ScriptDependencyContextReader(logFactory),
                 logFactory
             )
-        { }
-
-        private static IRestorer CreateRestorer(LogFactory logFactory)
         {
-            var commandRunner = new CommandRunner(logFactory);
-            return new ProfiledRestorer(new DotnetRestorer(commandRunner, logFactory),logFactory);
-        }
-
-        [Obsolete("Please use the overload that takes in a list of script files")]
-        public IEnumerable<CompilationDependency> GetDependencies(string targetDirectory, bool enableScriptNugetReferences, string defaultTargetFramework = "net46")
-        {
-            return GetDependencies(targetDirectory, null, enableScriptNugetReferences, defaultTargetFramework);
         }
 
         public IEnumerable<CompilationDependency> GetDependencies(string targetDirectory, IEnumerable<string> scriptFiles, bool enableScriptNugetReferences, string defaultTargetFramework = "net46")
         {
-            var pathToProjectFile = scriptFiles == null ?
-                _scriptProjectProvider.CreateProject(targetDirectory, defaultTargetFramework, enableScriptNugetReferences) :
-                _scriptProjectProvider.CreateProject(targetDirectory, scriptFiles, defaultTargetFramework, enableScriptNugetReferences);
-
-            if (pathToProjectFile == null)
+            var pathToProjectFile = _scriptProjectProvider.CreateProject(targetDirectory, scriptFiles,defaultTargetFramework, enableScriptNugetReferences);
+            _restorer.Restore(pathToProjectFile, packageSources: Array.Empty<string>());
+            var pathToAssetsFile = Path.Combine(Path.GetDirectoryName(pathToProjectFile), "obj", "project.assets.json");
+            var dependencyContext = _scriptDependencyContextReader.ReadDependencyContext(pathToAssetsFile);
+            var result = new List<CompilationDependency>();
+            foreach (var scriptDependency in dependencyContext.Dependencies)
             {
-                return Array.Empty<CompilationDependency>();
-            }
-
-            var dependencyInfo = _scriptDependencyInfoProvider.GetDependencyInfo(pathToProjectFile, Array.Empty<string>());
-            var dependencyContext = dependencyInfo.DependencyContext;
-            var compilationAssemblyResolvers = GetCompilationAssemblyResolvers(dependencyInfo.NugetPackageFolders);
-
-            List<CompilationDependency> result = new List<CompilationDependency>();
-            var compileLibraries = dependencyContext.CompileLibraries;
-
-            foreach (var compilationLibrary in compileLibraries)
-            {
-                var resolvedReferencePaths = new HashSet<string>();
-                _logger.Trace($"Resolving compilation reference paths for {compilationLibrary.Name}");
-                var referencePaths = ResolveReferencePaths(compilationLibrary, compilationAssemblyResolvers);
-                var userPackageFolder = dependencyInfo.NugetPackageFolders.First();
-                var fallbackFolders = dependencyInfo.NugetPackageFolders.Skip(1);
-                var packagePathResolver = new FallbackPackagePathResolver(userPackageFolder, fallbackFolders);
-
-                var packagePath = packagePathResolver.GetPackageDirectory(compilationLibrary.Name, compilationLibrary.Version);
-
-                var scripts =
-                    _scriptFilesDependencyResolver.GetScriptFileDependencies(packagePath);
-                foreach (var referencePath in referencePaths)
-                {
-                    resolvedReferencePaths.Add(referencePath);
-                }
-                var compilationDependency = new CompilationDependency(
-                    compilationLibrary.Name,
-                    compilationLibrary.Version,
-                    resolvedReferencePaths.ToList(),
-                    scripts);
-
+                var compilationDependency = new CompilationDependency(scriptDependency.Name, scriptDependency.Version, scriptDependency.CompileTimeDependencyPaths, scriptDependency.ScriptPaths);
                 result.Add(compilationDependency);
             }
             return result;
         }
 
-        private IEnumerable<string> ResolveReferencePaths(CompilationLibrary compilationLibrary, ICompilationAssemblyResolver[] compilationAssemblyResolvers)
+        private static IRestorer CreateRestorer(LogFactory logFactory)
         {
-            if (compilationLibrary.Assemblies.Any(a => a.EndsWith("_._")))
-            {
-                return Array.Empty<string>();
-            }
-            var referencePaths = compilationLibrary.ResolveReferencePaths(compilationAssemblyResolvers).Select(p => Path.GetFullPath(p)).ToArray();
-
-            foreach (var referencePath in referencePaths)
-            {
-                _logger.Trace($"{compilationLibrary.Name} => {referencePath}");
-            }
-
-            return referencePaths;
-        }
-
-        private ICompilationAssemblyResolver[] GetCompilationAssemblyResolvers(string[] nugetPackageFolders)
-        {
-            var resolvers = new List<ICompilationAssemblyResolver>();
-
-
-            foreach (var nugetPackageFolder in nugetPackageFolders)
-            {
-                resolvers.Add(CreatePackageResolver(nugetPackageFolder));
-            }
-            return resolvers.ToArray();
-        }
-
-        private PackageCompilationAssemblyResolver CreatePackageResolver(string nugetGetPackageDirectory)
-        {
-            _logger.Debug($"Creating {nameof(PackageCompilationAssemblyResolver)} for target path: {nugetGetPackageDirectory}");
-            return new PackageCompilationAssemblyResolver(nugetGetPackageDirectory);
-        }
-    }
-
-    public class CompilationDependency
-    {
-        public CompilationDependency(string name, string version, IReadOnlyList<string> assemblyPaths, IReadOnlyList<string> scripts)
-        {
-            Name = name;
-            Version = version;
-            AssemblyPaths = assemblyPaths;
-            Scripts = scripts;
-        }
-
-        public string Name { get; }
-
-        public string Version { get; }
-
-        public IReadOnlyList<string> AssemblyPaths { get; }
-
-        public IReadOnlyList<string> Scripts { get; }
-
-        public override string ToString()
-        {
-            return $"Name: {Name} , Version: {Version}";
+            var commandRunner = new CommandRunner(logFactory);
+            return new ProfiledRestorer(new DotnetRestorer(commandRunner, logFactory),logFactory);
         }
     }
 }
