+++ src/Dotnet.Script.DependencyModel/Dotnet.Script.DependencyModel.csproj
@@ -24,7 +24,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyModel" Version="2.1.0" />
     <PackageReference Include="NuGet.ProjectModel" Version="4.9.3" />
   </ItemGroup>
