+++ src/Dotnet.Script.Tests/CompilationDependencyResolverTests.cs
@@ -9,7 +9,7 @@ namespace Dotnet.Script.Tests
 {
     [Collection("IntegrationTests")]
     public class CompilationDependencyResolverTests
-    {        
+    {
         private readonly ScriptEnvironment _scriptEnvironment;
 
         public CompilationDependencyResolverTests(ITestOutputHelper testOutputHelper)
@@ -20,7 +20,7 @@ namespace Dotnet.Script.Tests
 
         [Fact]
         public void ShouldGetCompilationDependenciesForPackageContainingInlineNuGetPackageReference()
-        {            
+        {
             var resolver = CreateResolver();
             var targetDirectory = TestPathUtils.GetPathToTestFixtureFolder("InlineNugetPackage");
             var csxFiles = Directory.GetFiles(targetDirectory, "*.csx");
@@ -46,7 +46,7 @@ namespace Dotnet.Script.Tests
             var targetDirectory = TestPathUtils.GetPathToTestFixtureFolder("NativeLibrary");
             var csxFiles = Directory.GetFiles(targetDirectory, "*.csx");
             var dependencies = resolver.GetDependencies(targetDirectory, csxFiles, true, _scriptEnvironment.TargetFramework);
-            Assert.Contains(dependencies, d => d.Name == "Microsoft.Data.Sqlite");
+            Assert.Contains(dependencies, d => d.Name == "Microsoft.Data.Sqlite.Core");
         }
 
         [Fact]
@@ -57,12 +57,12 @@ namespace Dotnet.Script.Tests
             var csxFiles = Directory.GetFiles(targetDirectory, "*.csx");
             var dependencies = resolver.GetDependencies(targetDirectory, csxFiles, true, _scriptEnvironment.TargetFramework);
             Assert.Contains(dependencies, d => d.Name == "Auth0.ManagementApi");
-        }       
+        }
 
         private CompilationDependencyResolver CreateResolver()
         {
             var resolver = new CompilationDependencyResolver(TestOutputHelper.CreateTestLogFactory());
-            return resolver;            
+            return resolver;
         }
     }
 }
